refactor(exec): reorganize
- Rename `loader` module to `runtime`
- Increase clarity and readability of code by keeping related
  functionality close:
  - introduce `workload` module
  - introduce `identity` module
  - introduce `io` module
  - introduce `net` module
- Increase test coverage
- Simplify the structure and improve readability and code navigation by removing the state machine

  This also ensures we have an organization, which is better suited for
  upcoming features being added to the crate (e.g. VFS)

  While the state machine abstraction is technically an interesting one,
  there are a few drawbacks hurting development associated, mostly due to
  the fact that the abstraction is applied too early in the lifetime of the
  project and the finite (or, at least, MVP) set of features and use cases
  covered is not clear yet and is expected to change rapidly.
  - High degree of abstraction hurts development iteration speed
     and collaboration on the project, due to sheer amount of moving
     pieces that have to be changed to implement new features.
  - Unconventional structure makes onboarding developers and getting them
     up-to-speed unnecessarily harder (sample size of 1).
  - Arguably, the state machine is the wrong abstraction for this use
     case, which results in workarounds having to be put in place to
     address the assocated fragmentation of implementation.
